app/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
python-dotenv==1.0.1
httpx==0.26.0
openai==1.12.0
deepgram-sdk==3.1.6
aiohttp==3.9.3
